chore: ProjectStatusTag 공용 컴포넌트 위치 이동
